// examples/uss-omaha_batch-4/uss-omaha/src/types/content.ts
export interface Metadata {
  title: string
  subtitle: string
  year: string
  mode: 'memorial' | 'donor'
}
// ...
export interface CallToActionSection {
  memorial: CallToActionMode
  donor: CallToActionMode
}

export interface CallToActionMode {
  heading: string
  tagline: string
  donationHeading: string
  primaryOrg: Organization
  alternateOrg: AlternateOrganization
  taxNote: string
  pledgeFormText: string
  pledgeFormUrl: string
}

export interface Organization {
  name: string
  ein: string
  website: string
  email: string
  mailingAddress: MailingAddress
  phone: string
}

export interface AlternateOrganization {
  name: string
  ein: string
  note: string
}